AI工作時考慮資金糧草，補充所有部隊

+++ GameObjects/GameObjects/Troop.cs
@@ -1760,16 +1760,27 @@
                                 {
                                     // 去除多余的点
                                     int i = 0;
+                                    bool skip = false;
                                     while (startingArea.HasPoint(this.FirstTierPath[i]))
+                                    {
                                         i++;
-                                    this.FirstTierPath.RemoveRange(0, i);
-                                    i = this.FirstTierPath.Count - 1;
-                                    while (willArea.HasPoint(this.FirstTierPath[i]))
-                                        i--;
-                                    this.FirstTierPath.RemoveRange(i + 1, this.FirstTierPath.Count - i - 1);
+                                        if (i >= this.FirstTierPath.Count)
+                                        {
+                                            skip = true;
+                                            break;
+                                        }
+                                    }
+                                    if (!skip)
+                                    {
+                                        this.FirstTierPath.RemoveRange(0, i);
+                                        i = this.FirstTierPath.Count - 1;
+                                        while (willArea.HasPoint(this.FirstTierPath[i]))
+                                            i--;
+                                        this.FirstTierPath.RemoveRange(i + 1, this.FirstTierPath.Count - i - 1);
 
-                                    base.Scenario.pathCache[new PathCacheKey(this.StartingArchitecture, this.WillArchitecture, this.Army.Kind)] = this.FirstTierPath;
-                                    path = ConstructTruePath(this.FirstTierPath, kind);
+                                        base.Scenario.pathCache[new PathCacheKey(this.StartingArchitecture, this.WillArchitecture, this.Army.Kind)] = this.FirstTierPath;
+                                        path = ConstructTruePath(this.FirstTierPath, kind);
+                                    }
                                 }
                                 else
                                 {
